+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## [3.5.3] - 2024-11-17
 
 ### Added
-- `config generate` API for generating a configuration file that is synchronized with its source file.
-- Initial implementation of remote execution server in lock mode, preventing parallel monorail use for a subset of APIs
+- `config generate` API for generating a configuration file that is synchronized with its source file
+- Initial implementation of lock server, preventing parallel monorail use for a subset of APIs
 
 ## [3.5.2] - 2024-11-15